jakarta_ecoguard/requirements.txt
python-dotenv
crewai
langchain-google-genai
langchain-openai
requests